Split build pipeline into clean, bundle, and types steps
