+++ src/cli.ts
@@ -1,0 +1,23 @@
+import { start_repl } from './repl';
+
+const banner = `
+
+░█████╗░░█████╗░██████╗░███████╗██╗░░██╗██╗██╗░░░██╗░█████╗░
+██╔══██╗██╔══██╗██╔══██╗██╔════╝╚██╗██╔╝██║██║░░░██║██╔══██╗
+██║░░╚═╝██║░░██║██║░░██║█████╗░░░╚███╔╝░██║╚██╗░██╔╝██║░░██║
+██║░░██╗██║░░██║██║░░██║██╔══╝░░░██╔██╗░██║░╚████╔╝░██║░░██║
+╚█████╔╝╚█████╔╝██████╔╝███████╗██╔╝╚██╗██║░░╚██╔╝░░╚█████╔╝
+░╚════╝░░╚════╝░╚═════╝░╚══════╝╚═╝░░╚═╝╚═╝░░░╚═╝░░░░╚════╝░
+`;
+
+function main() {
+  console.log('Bienvenido al REPL de: ');
+  console.log(banner);
+  start_repl();
+}
+
+if (typeof require !== 'undefined' && require.main === module) {
+  main();
+}
+
+export { banner, main };
